requirements.txt
absl-py==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
branca==0.5.0
cachetools==5.0.0
catboost==1.0.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-admin-thumbnails==0.2.6
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-decouple==2.1
django-embed-video==1.4.4
django-filter==22.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
dominate==2.6.0
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
flatbuffers==2.0
folium==0.12.1.post1
fonttools==4.30.0
future==0.18.2
gast==0.5.3
geocoder==1.38.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
graphviz==0.19.1
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
multimethod==1.7
mysqlclient==2.1.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.8
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.8.3
phik==0.12.1
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
plotly==5.6.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
ratelim==0.1.6
reportlab==3.6.10
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.3
visions==0.7.4
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==6.0.0
widgetsnbextension==3.6.0
wrapt==1.14.0
xgboost==1.5.2
